// src/generated-types/store-category.ts
export interface Id {
  id: string;
}

/** Message representing a status response */
export interface StatusResponse {
  success: boolean;
  message: string;
}

/** Message representing a list of store categories */
export interface StoreCategoryList {
  data: StoreCategoryWithTranslation[];
}

/** Message representing a store category */
export interface StoreCategory {
  id: string;
  slug: string;
  isAvailable: boolean;
  sortOrder: number;
}

/** Message representing a store category with translation */
export interface StoreCategoryWithTranslation {
  id: string;
  slug: string;
  isAvailable: boolean;
  sortOrder: number;
  title: string;
  description: string;
}

/** Message representing a store category translation */
export interface StoreCategoryTranslation {
  id: string;
  title: string;
  description: string;
  language: Language1;
}

/** Message representing a store category with its translation */
export interface StoreCategoryWithTranslations {
  category: StoreCategory | null;
  translation: StoreCategoryTranslation[];
}

/** Message for creating a new store category */
export interface CreateStoreCategoryRequest {
  slug: string;
  isAvailable?: boolean | null | undefined;
  sortOrder?: number | null | undefined;
}

/** Message for updating an existing store category */
export interface UpdateStoreCategoryRequest {
  id: string;
  slug?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
}

/** Message for changing the position of a store category */
export interface ChangeStoreCategoryPositionRequest {
  id: string;
  sortOrder: number;
}